Escape ! by escaping it with a preceding \
! is a reserved character in the MarkdownV2-style of Telegram and must
therefore be preceded with \ as described in their document.

From the Telegram API documentation:
Any character with code between 1 and 126 inclusively can be escaped
anywhere with a preceding '\' character, in which case it is treated as
an ordinary character and not a part of the markup. This implies that '\'
character usually must be escaped with a preceding '\' character.

Telegram API documentation:
https://core.telegram.org/bots/api#markdownv2-style

Fixes https://github.com/ikelax/mensa-bot/issues/8