Show inline query results for all days
So far, only an inline query result for today was shown. Now, for each
day provided by the API an inline query result is shown.

The dates are sorted such that dates in the future are at the top of the
list in ascending order and dates in the past are at the bottom in
descending order.

Also, the replied mealplan contains the date of the mealplan as a title
now.

Implements https://github.com/ikelax/mensa-bot/issues/7