Escape brackets in meal plan title in Markdown
If the meal plan is for today or for a date in past, (heute) or
(vergangen) is appended to it. The bot sends the meal plan in the
Markdown style of Telegram. There, brackets are reserved characters and
must therefore be escaped with \\. In the inline queries, the brackets
are not escaped because Markdown is not used there.

Closes https://github.com/ikelax/mensa-bot/issues/34